Added upnp port mapping view along with port mapping removal.
